Protocol refactoring fix, relation mapping fix, upgrade Engine @4.68.0, SDLC @0.179.1
