+++ legend-engine-ide-lsp-default-extensions/src/main/java/org/finos/legend/engine/ide/lsp/extension/mapping/MappingLSPGrammarExtension.java
@@ -47,7 +47,7 @@ import org.finos.legend.engine.plan.generation.transformers.PlanTransformer;
 import org.finos.legend.engine.protocol.analytics.model.MappedEntity;
 import org.finos.legend.engine.protocol.analytics.model.MappingModelCoverageAnalysisResult;
 import org.finos.legend.engine.protocol.pure.v1.model.context.PureModelContextData;
-import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.PackageableElement;
+import org.finos.legend.engine.protocol.pure.v1.model.PackageableElement;
 import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.mapping.*;
 import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.mapping.aggregationAware.AggregationAwareClassMapping;
 import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.mapping.aggregationAware.AggregationAwarePropertyMapping;
@@ -59,6 +59,8 @@ import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.mapping
 import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.mapping.xStore.XStorePropertyMapping;
 import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.store.modelToModel.mapping.PureInstanceClassMapping;
 import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.store.modelToModel.mapping.PurePropertyMapping;
+import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.mapping.relationFunction.RelationFunctionClassMapping;
+import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.mapping.relationFunction.RelationFunctionPropertyMapping;
 import org.finos.legend.engine.protocol.pure.v1.model.test.AtomicTest;
 import org.finos.legend.engine.protocol.pure.v1.model.test.TestSuite;
 import org.finos.legend.engine.pure.code.core.PureCoreExtensionLoader;
@@ -407,6 +409,12 @@ public class MappingLSPGrammarExtension extends AbstractLegacyParserLSPGrammarEx
             {
                 return Stream.empty();
             }
+
+            @Override
+            public Stream<Optional<LegendReferenceResolver>> visit(RelationFunctionClassMapping relationFunctionClassMapping)
+            {
+                return Stream.empty();
+            }
         });
 
         return Stream.concat(Stream.of(legendReferenceResolver), otherReferences);
@@ -456,6 +464,12 @@ public class MappingLSPGrammarExtension extends AbstractLegacyParserLSPGrammarEx
 
             @Override
             public Stream<Optional<LegendReferenceResolver>> visit(XStorePropertyMapping xStorePropertyMapping)
+            {
+                return Stream.empty();
+            }
+
+            @Override
+            public Stream<Optional<LegendReferenceResolver>> visit(RelationFunctionPropertyMapping aggregationAwarePropertyMapping)
             {
                 return Stream.empty();
             }
