+++ pom.xml
@@ -49,8 +49,8 @@
     </modules>
 
     <properties>
-        <legend.engine.version>4.67.5</legend.engine.version>
-        <legend.sdlc.version>0.178.2</legend.sdlc.version>
+        <legend.engine.version>4.68.0</legend.engine.version>
+        <legend.sdlc.version>0.179.1</legend.sdlc.version>
 
         <logback.version>1.4.14</logback.version>
